+++ library/src/main/java/de/nenick/espressomacchiato/elements/EspSupportAlertDialog.java
@@ -1,28 +1,26 @@
 package de.nenick.espressomacchiato.elements;
 
-import de.nenick.espressomacchiato.tools.EspResourceTool;
-
 /**
  * Pre configured {@link EspDialog} for general {@link android.support.v7.app.AlertDialog}.
  */
-public class EspAlertDialog extends EspDialog {
+public class EspSupportAlertDialog extends EspDialog {
 
     @Deprecated // mark parent static method as not usable for this class
     public static Spec spec() {
         throw new UnsupportedOperationException();
     }
 
-    public static EspAlertDialog build() {
-        return new EspAlertDialog(EspDialog.spec()
-                .withRoot(EspResourceTool.idByName("parentPanel"))
-                .withTitle(EspResourceTool.idByName("alertTitle"))
+    public static EspSupportAlertDialog build() {
+        return new EspSupportAlertDialog(EspDialog.spec()
+                .withRoot(android.support.design.R.id.parentPanel)
+                .withTitle(android.support.design.R.id.alertTitle)
                 .withMessage(android.R.id.message)
                 .withConfirmButton(android.R.id.button1)
                 .withDenyButton(android.R.id.button2)
                 .withCancelButton(android.R.id.button3));
     }
 
-    public EspAlertDialog(Spec spec) {
+    public EspSupportAlertDialog(Spec spec) {
         super(spec);
     }
 }